MLK-23726 drivers: caam: Fix CAAM resume function for i.MX 6SX
i.MX6SX low power mode removes CAAM power. The order in which the
registers are restored is important for JR configuration.
The MID registers need to be configured first in order for the other
ones to be configured correctly. This commit makes a call to the backup
function to save the MID settings for the JR used by OP-TEE first.

Fixes: 65d11b9078be ("TEE-511 drivers: caam: Fix suspend/resume")
Signed-off-by: Silvano di Ninno <silvano.dininno@nxp.com>
(cherry picked from commit 83fd4bc4037bebdbdfe50389edf58c6c1bb04881)