LFOPTEE-9 plat-ls: Increase heap size
The test xtest 1001 can fail because of internal mechanism
happening more quickly when the heap is smaller. In this
case the failure does not indicate a bug but that the test
did not take account a precise behavior of the memory
allocator.

Signed-off-by: Franck LENORMAND <franck.lenormand@nxp.com>
(cherry picked from commit 63fc2437cba3891ccbb61a49c93eb2c56aea35bb)